Extract footer in a component and create the welcome modal
