+++ src/components/App.js
@@ -3,18 +3,14 @@ import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { BrowserRouter as Router } from 'react-router-dom';
 import secretin from 'utils/secretin';
-import Secretin from 'secretin';
 
 import * as AppUIActions from 'slices/AppUISlice';
 import { getSecretsInFolder } from 'selectors/MetadataSelectors';
-import Modal from 'components/utilities/Modal';
-import Button from 'components/utilities/Button';
+import Footer from 'components/Footer';
 
 import UserConnect from 'components/users/UserConnect';
 import Layout from 'components/Layout';
 
-const LOCAL_STORAGE_ACKNOWLEDGE_VERSION_KEY = 'Secret-in:acknowledgedVersion';
-const DEFAULT_DEVELOP_COMMITISH = 'develop';
 class App extends Component {
   static propTypes = {
     savedUsername: PropTypes.string,
@@ -32,23 +28,6 @@ class App extends Component {
     this.onAppFocus = this.onAppFocus.bind(this);
     this.onAppBlur = this.onAppBlur.bind(this);
     this.handleConnectionChange = this.handleConnectionChange.bind(this);
-    this.hideNews = this.hideNews.bind(this);
-    this.showNews = this.showNews.bind(this);
-    this.acknowledgeVersion = this.acknowledgeVersion.bind(this);
-
-    // eslint-disable-next-line no-undef
-    const commit = SECRETIN_APP_COMMIT;
-
-    const acknowledgedVersion = window.localStorage.getItem(
-      LOCAL_STORAGE_ACKNOWLEDGE_VERSION_KEY
-    );
-
-    this.state = {
-      commit,
-      acknowledgedVersion,
-      news: [],
-      showNews: false,
-    };
 
     window.addEventListener('focus', this.onAppFocus);
     window.addEventListener('blur', this.onAppBlur);
@@ -56,38 +35,6 @@ class App extends Component {
 
   componentDidMount() {
     secretin.addEventListener('connectionChange', this.handleConnectionChange);
-
-    const somethingNew =
-      this.state.commit !== DEFAULT_DEVELOP_COMMITISH &&
-      this.state.acknowledgedVersion !== null &&
-      this.state.acknowledgedVersion !== this.state.commit;
-    if (somethingNew) {
-      fetch(
-        'https://raw.githubusercontent.com/secretin/secretin-website/develop/changelog.json'
-      )
-        .then(response => {
-          return response.json();
-        })
-        .then(data => {
-          const news = [];
-          let foundAcknowledgedVersion = false;
-          for (const change of data.changelog) {
-            if (change.version === this.state.acknowledgedVersion) {
-              foundAcknowledgedVersion = true;
-              break;
-            }
-            news.push(change);
-          }
-
-          if (foundAcknowledgedVersion) {
-            this.setState({ news });
-          } else {
-            // If we didn't find the acknowledged version, then we didn't updated changelog properly
-            // we will just show the last change then
-            this.setState({ news: [data.changelog[0]] });
-          }
-        });
-    }
   }
 
   componentWillUnmount() {
@@ -103,22 +50,6 @@ class App extends Component {
 
   onAppFocus() {
     clearTimeout(this.disconnectingEvent);
-  }
-
-  showNews() {
-    this.setState({ showNews: true });
-  }
-
-  hideNews() {
-    this.setState({ showNews: false });
-  }
-
-  acknowledgeVersion() {
-    window.localStorage.setItem(
-      LOCAL_STORAGE_ACKNOWLEDGE_VERSION_KEY,
-      this.state.commit
-    );
-    this.setState({ news: [] });
   }
 
   onAppBlur() {
@@ -137,57 +68,6 @@ class App extends Component {
   }
 
   render() {
-    console.log(this.state.news);
-    const shortCommit = this.state.commit ? this.state.commit.substr(0, 7) : '';
-    const secretinAppVersion = this.state.commit ? (
-      <span className="secretin-version">
-        {this.state.news.length > 0 && (
-          <>
-            <Modal show={this.state.showNews} onClose={this.hideNews}>
-              <Modal.Header title="What's new ?" />
-              <Modal.Body>
-                {this.state.news.map(change => (
-                  <div key={change.version}>
-                    <h2>{change.title}</h2>
-                    <p>{change.description}</p>
-                  </div>
-                ))}
-              </Modal.Body>
-
-              <Modal.Footer>
-                <Button
-                  type="reset"
-                  buttonStyle="default"
-                  onClick={this.hideNews}
-                >
-                  Close
-                </Button>
-                <Button type="submit" onClick={this.acknowledgeVersion}>
-                  Ok
-                </Button>
-              </Modal.Footer>
-            </Modal>
-            <span className="new-in-secretin-version" onClick={this.showNews}>
-              •
-            </span>
-          </>
-        )}
-        <span className="secretin-version-text">
-          secretin-app{' '}
-          <a
-            href={`https://github.com/secretin/secretin-app/commit/${this.state.commit}`}
-            rel="noopener noreferrer"
-            target="_blank"
-          >
-            {shortCommit}
-          </a>
-        </span>
-      </span>
-    ) : (
-      <span className="secretin-version">
-        <span className="secretin-version-text">secretin-app dev</span>
-      </span>
-    );
     return (
       <Router basename={process.env.PUBLIC_URL}>
         <div className="App">
@@ -200,14 +80,7 @@ class App extends Component {
               errors={this.props.errors}
             />
           )}
-          <div className="footer">
-            {secretinAppVersion}
-            <span className="secretin-version">
-              <span className="secretin-version-text">
-                secretin-lib v{Secretin.version}
-              </span>
-            </span>
-          </div>
+          <Footer />
         </div>
       </Router>
     );
